+++ x/btcstaking/keeper/keeper.go
@@ -66,81 +66,8 @@ func (k Keeper) Logger(ctx sdk.Context) log.Logger {
 func (k Keeper) BeginBlocker(ctx context.Context) error {
 	// index BTC height at the current height
 	k.IndexBTCHeight(ctx)
-
-	params := k.GetParams(ctx)
-	btcTipHeight, err := k.GetCurrentBTCHeight(ctx)
-	if err != nil {
-		panic(err) // only possible upon programming error
-	}
-	wValue := k.btccKeeper.GetParams(ctx).CheckpointFinalizationTimeout
-
-	// prepare metrics for {active, inactive} finality providers,
-	// {pending, active, unbonded BTC delegations}, and total staked Bitcoins
-	// NOTE: slashed finality providers and BTC delegations are recorded upon
-	// slashing events rather than here
-	var (
-		numFPs        int    = 0
-		numStakedSats uint64 = 0
-		numDelsMap           = map[types.BTCDelegationStatus]int{
-			types.BTCDelegationStatus_PENDING:  0,
-			types.BTCDelegationStatus_ACTIVE:   0,
-			types.BTCDelegationStatus_UNBONDED: 0,
-		}
-	)
-	// for voting power and rewards
-	dc := types.NewVotingPowerDistCache()
-
-	// iterate over all finality providers to find out non-slashed ones that have
-	// positive voting power
-	k.IterateActiveFPs(
-		ctx,
-		func(fp *types.FinalityProvider) bool {
-			fpDistInfo := types.NewFinalityProviderDistInfo(fp)
-
-			// iterate over all BTC delegations under the finality provider
-			// in order to accumulate voting power dist info for it
-			k.IterateBTCDelegations(ctx, fp.BtcPk, func(btcDel *types.BTCDelegation) bool {
-				// accumulate voting power and reward distribution cache
-				fpDistInfo.AddBTCDel(btcDel, btcTipHeight, wValue, params.CovenantQuorum)
-
-				// record metrics
-				numStakedSats += btcDel.VotingPower(btcTipHeight, wValue, params.CovenantQuorum)
-				numDelsMap[btcDel.GetStatus(btcTipHeight, wValue, params.CovenantQuorum)]++
-
-				return true
-			})
-
-			if fpDistInfo.TotalVotingPower > 0 {
-				dc.AddFinalityProviderDistInfo(fpDistInfo)
-			}
-
-			return true
-		},
-	)
-	// record metrics for finality providers and total staked BTCs
-	numActiveFPs := min(numFPs, int(params.MaxActiveFinalityProviders))
-	types.RecordActiveFinalityProviders(numActiveFPs)
-	types.RecordInactiveFinalityProviders(numFPs - numActiveFPs)
-	numStakedBTCs := float32(numStakedSats / SatoshisPerBTC)
-	types.RecordMetricsKeyStakedBitcoins(numStakedBTCs)
-	// record metrics for BTC delegations
-	for status, num := range numDelsMap {
-		types.RecordBTCDelegations(num, status)
-	}
-
-	// filter out top `MaxActiveFinalityProviders` active finality providers in terms of voting power
-	maxNumActiveFPs := k.GetParams(ctx).MaxActiveFinalityProviders
-	activeFps := types.FilterTopNFinalityProviders(dc.FinalityProviders, maxNumActiveFPs)
-	// set voting power table and re-calculate total voting power of top N finality providers
-	dc.TotalVotingPower = uint64(0)
-	babylonTipHeight := uint64(sdk.UnwrapSDKContext(ctx).HeaderInfo().Height)
-	for _, fp := range activeFps {
-		k.SetVotingPower(ctx, fp.BtcPk.MustMarshal(), babylonTipHeight, fp.TotalVotingPower)
-		dc.TotalVotingPower += fp.TotalVotingPower
-	}
-
-	// set the voting power distribution cache of the current height
-	k.setVotingPowerDistCache(ctx, uint64(sdk.UnwrapSDKContext(ctx).HeaderInfo().Height), dc)
+	// update voting power distribution
+	k.UpdatePowerDist(ctx)
 
 	return nil
 }
